requirements.txt
telethon>=1.27
pandas>=1.0
python-dotenv>=1.0
openpyxl
apify-client>=1.6.0
aiohttp>=3.9.0
truststore>=0.9.0
